bump nodeset filter limit to 300 (from 200)
Cloudflare was throwing a 'quota exceeded' error.
We reduced the limit and manually deleted TXT entries.
This has succeeded in making Cloudflare not error.
With a limit of 200, CF currently has 480 records of a
max 1000. This patch tries to nudge the number of records
a little higher, without going too high.